lesson5 updated CRUD with database

+++ src/services/user.service.ts
@@ -1,3 +1,4 @@
+import { ApiError } from "../api-error";
 import { userRepository } from "../repositories/user.repository";
 import { IUser } from "../user.interface";
 
@@ -8,13 +9,25 @@ class UserService {
   public async createUser(dto: Partial<IUser>): Promise<IUser> {
     return await userRepository.createUser(dto);
   }
-  public async getById(id: number): Promise<IUser> {
-    return await userRepository.getById(id);
+  public async getById(id: string): Promise<IUser> {
+    const user = await userRepository.getById(id);
+    if (!user) {
+      throw new ApiError("user not found", 404);
+    }
+    return user;
   }
-  public async putById(data: Partial<IUser>, id: number) {
+  public async putById(data: Partial<IUser>, id: string) {
+    const user = await userRepository.getById(id);
+    if (!user) {
+      throw new ApiError("user not found", 404);
+    }
     return await userRepository.putById(data, id);
   }
-  public async deleteById(id: number): Promise<IUser[]> {
+  public async deleteById(id: string): Promise<IUser[]> {
+    const user = await userRepository.getById(id);
+    if (!user) {
+      throw new ApiError("user not found", 404);
+    }
     return await userRepository.deleteById(id);
   }
 }
